core/ai_support_bot/pyproject.toml
[project]
name = "ai_support_bot"
version = "0.1.0"
description = "24/7 Discord AI Customer Support Bot with RAG (Notion + Google Sheets + Gemini)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "discord.py>=2.3.0",
    "openai>=1.0.0",
    "notion-client>=2.2.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
    "pythainlp>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (may need network)",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
